Add LIC 6 test case
