+++ src/main/java/decide/LIC.java
@@ -1,6 +1,7 @@
 package decide;
 
 import java.awt.Point;
+import java.awt.geom.Point2D;
 import java.util.ArrayList;
 import java.util.HashSet;
 
@@ -15,7 +16,7 @@ public class LIC {
    * are a distance greater than the length, LENGTH1, apart.
    * (0 ≤ LENGTH1)
    *
-   * @param points Array containing the coordinates of data points
+   * @param points  Array containing the coordinates of data points
    * @param length1 Distance that two points need to be from each other
    * @return True if LIC 1 is met
    */
@@ -38,7 +39,7 @@ public class LIC {
    * cannot all be contained within or on a circle of radius RADIUS1.
    * (0 ≤ RADIUS1)
    *
-   * @param points Array containing the coordinates of data points.
+   * @param points  Array containing the coordinates of data points.
    * @param radius1 Radius of circle that should not cover the points.
    * @return True if LIC2 is met
    */
@@ -53,11 +54,9 @@ public class LIC {
       b = points.get(i + 1);
       c = points.get(i + 2);
       Point center = new Point((a.x + b.x + c.x) / 3, (a.y + b.y + c.y) / 3);
-      if (
-          distance(a, center) >= radius1
+      if (distance(a, center) >= radius1
           && distance(b, center) >= radius1
-          && distance(c, center) >= radius1
-      ) {
+          && distance(c, center) >= radius1) {
         return true;
       }
     }
@@ -83,8 +82,7 @@ public class LIC {
       b = points.get(i + 1);
       c = points.get(i + 2);
       area = Math.abs(
-              (a.x * (b.y - c.y) + b.x * (c.y - a.y) + c.x * (a.y - b.y))
-      ) / 2;
+          (a.x * (b.y - c.y) + b.x * (c.y - a.y) + c.x * (a.y - b.y))) / 2;
       if (area > area1) {
         return true;
       }
@@ -102,19 +100,18 @@ public class LIC {
    * is in quadrant III, the point (0,1) is in quadrant I and the point (1,0)
    * is in quadrant I. (2 ≤ Q PTS ≤ NUMPOINTS), (1 ≤ QUADS ≤ 3)
    *
-   * @param points Array containing the coordinates of data points
-   * @param qpts Number of consecutive points
-   * @param quads Least number of quadrants that should be inhabited minus one
+   * @param points    Array containing the coordinates of data points
+   * @param qpts      Number of consecutive points
+   * @param quads     Least number of quadrants that should be inhabited minus one
    * @param numpoints Number of points in points array
    * @return True if LIC 5 is met
    */
   @SuppressWarnings("checkstyle:magicnumber")
   protected boolean lic5(
-          final ArrayList<Point> points,
-          final int qpts,
-          final int quads,
-          final int numpoints
-  ) {
+      final ArrayList<Point> points,
+      final int qpts,
+      final int quads,
+      final int numpoints) {
     assert (qpts > 1 && qpts <= numpoints);
     assert (quads > 0 && quads < 4);
 
@@ -133,26 +130,77 @@ public class LIC {
   }
 
   /**
+   * There exists at least one set of N_PTS consecutive data points such that at
+   * least one of the points lies a distance greater than DIST from the line
+   * joining the first and last of these N_PTS points. If the first and last
+   * points of these N_PTS are identical, then the calculated distance to compare
+   * with DIST will be the distance from the coincident point to all other points
+   * of the N_PTS consecutive points. The condition is not met when NUMPOINTS < 3.
+   * (3 ≤ N_PTS ≤ NUMPOINTS), (0 ≤ DIST)
+   * 
+   * @param points    Array containing the coordinates of data points
+   * @param numpoints The number of planar data points
+   * @param npts      Number of consecutive points
+   * @param dist      Distance
+   * 
+   * @return true iff LIC 6 is satisfied
+   */
+  protected boolean lic6(final ArrayList<Point2D> points, final int numpoints, final int npts, final double dist) {
+
+    if (numpoints < 3)
+      return false;
+
+    if (npts > numpoints || npts < 3)
+      return false;
+
+    if (dist < 0)
+      return false;
+
+    for (int i = 0; i < numpoints - npts + 1; i++) {
+      Point2D firstPoint = points.get(i);
+      Point2D lastPoint = points.get(i + npts - 1);
+
+      boolean isLine = !firstPoint.equals(lastPoint);
+
+      if (isLine)
+        for (int j = i + 1; j < i + npts - 1; j++) {
+          double distance = pointToLineDistance(points.get(j), firstPoint, lastPoint);
+          if (Double.compare(distance, dist) == 1)
+            return true;
+        }
+      else
+        for (int j = i + 1; j < i + npts - 1; j++) {
+          double distance = pointToPointdistance(points.get(j), firstPoint);
+          if (Double.compare(distance, dist) == 1)
+            return true;
+        }
+
+    }
+
+    return false;
+
+  }
+
+  /**
    * There exists at least one set of three data points separated by exactly
    * E PTS and F PTS consecutive intervening points, respectively, that are
    * the vertices of a triangle with area greater
    * than AREA1. The condition is not met when NUMPOINTS &lt; 5.
    *
-   * @param points Array containing the coordinates of data points
+   * @param points    Array containing the coordinates of data points
    * @param numpoints The number of planar data points
-   * @param area1 Area in LICs
-   * @param epts Number of points between the 1st and the 2nd data point
-   * @param fpts Number of points between the 2nd and the 3rd data point
+   * @param area1     Area in LICs
+   * @param epts      Number of points between the 1st and the 2nd data point
+   * @param fpts      Number of points between the 2nd and the 3rd data point
    * @return true iff LIC 10 is met
    */
   @SuppressWarnings("checkstyle:magicnumber")
   protected boolean lic11(
-          final ArrayList<Point> points,
-          final int numpoints,
-          final int area1,
-          final int epts,
-          final int fpts
-  ) {
+      final ArrayList<Point> points,
+      final int numpoints,
+      final int area1,
+      final int epts,
+      final int fpts) {
     if (numpoints < 5) {
       return false;
     }
@@ -165,8 +213,7 @@ public class LIC {
       b = points.get(i + epts);
       c = points.get(i + epts + fpts);
       area = Math.abs(
-              (a.x * (b.y - c.y) + b.x * (c.y - a.y) + c.x * (a.y - b.y))
-      ) / 2;
+          (a.x * (b.y - c.y) + b.x * (c.y - a.y) + c.x * (a.y - b.y))) / 2;
       if (area > area1) {
         return true;
       }
@@ -185,23 +232,22 @@ public class LIC {
    * parts must be true for the LIC to be true. The condition is not met
    * when NUMPOINTS &lt; 5.
    *
-   * @param points Array containing the coordinates of data points
+   * @param points    Array containing the coordinates of data points
    * @param numpoints The number of planar data points
-   * @param area1 1st Area in LICs
-   * @param area2 2nd Area in LICs
-   * @param epts Number of points between the 1st and the 2nd data point
-   * @param fpts Number of points between the 2nd and the 3rd data point
+   * @param area1     1st Area in LICs
+   * @param area2     2nd Area in LICs
+   * @param epts      Number of points between the 1st and the 2nd data point
+   * @param fpts      Number of points between the 2nd and the 3rd data point
    * @return true iff LIC 14 is met
    */
   @SuppressWarnings("checkstyle:magicnumber")
   protected boolean lic15(
-          final ArrayList<Point> points,
-          final int numpoints,
-          final int area1,
-          final int area2,
-          final int epts,
-          final int fpts
-  ) {
+      final ArrayList<Point> points,
+      final int numpoints,
+      final int area1,
+      final int area2,
+      final int epts,
+      final int fpts) {
     if (numpoints < 5) {
       return false;
     }
@@ -216,8 +262,7 @@ public class LIC {
       b = points.get(i + epts);
       c = points.get(i + epts + fpts);
       area = Math.abs(
-              (a.x * (b.y - c.y) + b.x * (c.y - a.y) + c.x * (a.y - b.y))
-      ) / 2.0;
+          (a.x * (b.y - c.y) + b.x * (c.y - a.y) + c.x * (a.y - b.y))) / 2.0;
       if (area > area1) {
         isGreater = true;
       }
@@ -242,6 +287,7 @@ public class LIC {
 
   /**
    * Method to get which quadrant around the origin a point is.
+   * 
    * @param point Point
    * @return The quadrant as 1, 2, 3, or 4
    */
@@ -256,6 +302,43 @@ public class LIC {
     } else {
       return 4;
     }
+  }
+
+  /**
+   * Calculates the distance between a point and a line using Heron's formula.
+   * 
+   * @param point
+   * @param lineStart
+   * @param lineEnd
+   * @return the distance
+   */
+  private double pointToLineDistance(Point2D point, Point2D lineStart, Point2D lineEnd) {
+    double i = pointToPointdistance(point, lineStart);
+    double j = pointToPointdistance(lineStart, lineEnd);
+    double k = pointToPointdistance(point, lineEnd);
+
+    double s = (i + j + k) / 2;
+    double area = Math.sqrt(s * (s - i) * (s - j) * (s - k));
+
+    double distance = (2 * area) / j;
+    return distance;
+  }
+
+  /**
+   * Calculates the distance between two points.
+   * 
+   * @param point1
+   * @param point2
+   * @return the distance
+   */
+  private double pointToPointdistance(Point2D point1, Point2D point2) {
+    double x1 = point1.getX(), y1 = point1.getY();
+    double x2 = point2.getX(), y2 = point2.getY();
+
+    double i = x2 - x1;
+    double j = y2 - y1;
+
+    return Math.sqrt(Math.pow(i, 2) + Math.pow(j, 2));
   }
 
   /**
